refactor: move jsonrpc serde raw value to types
