+++ types/src/lib.rs
@@ -8,6 +8,9 @@ extern crate alloc;
 /// JSON-RPC 2.0 specification related types.
 pub mod jsonrpc;
 
+/// JSON-RPC 2.0 specification related types v2.
+pub mod jsonrpc_v2;
+
 /// Shared error type.
 pub mod error;
